Move options struct to read module
